+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # 0.3.17 [unreleased]
 - refactor: Switch bitswap implementation [PR 66]
+- feat: Added Ipfs::pubsub_events to receive subscribe and unsubscribe events to a subscribed topic [PR 70]
+- feat: Implement RepoProvider [PR 69]
 
 [PR 66]: https://github.com/dariusc93/rust-ipfs/pull/66
+[PR 69]: https://github.com/dariusc93/rust-ipfs/pull/69
+[PR 70]: https://github.com/dariusc93/rust-ipfs/pull/70
 
 # 0.3.16
 - fix: Return events from gossipsub stream [PR 68]